fix: complete MSVC compatibility overhaul (v1.9.4-alpha)
- Remove __restrict__ from Statevector member declarations: MSVC doesn't support
  restrict hints on class members, only on local pointers (which remain intact)
- Add QPP_POPCOUNT64 macro: maps __builtin_popcountll (GCC/Clang) to __popcnt64
  (MSVC) for population count intrinsic in expectation value calculations
- Fix OpenMP loop variable types: change size_t to int in #pragma omp parallel
  for loops (MSVC requires signed integral types). Applied across gates and ops:
  two_qubit.cpp, multi_qubit.cpp, operators.cpp, estimator.cpp
- Wrap #pragma omp simd reduction with #if !defined(_MSC_VER) in operators.cpp
  (MSVC OpenMP 2.0 legacy doesn't support simd pragma without -openmp:experimental)

Performance impact: zero. All hot-path restrict aliasing hints on local pointers
remain unchanged. SIMD loop hints on GCC/Clang unaffected.

Fixes MSVC compilation while maintaining performance on GCC/Clang.

+++ src/primitives/estimator.cpp
@@ -25,7 +25,7 @@ std::vector<double> Estimator::run_batch(
 
     // run_single is thread-safe: all state is local (bound_circuit, sim, result).
     #pragma omp parallel for schedule(dynamic, 1)
-    for (size_t i = 0; i < n; ++i) {
+    for (int i = 0; i < static_cast<int>(n); ++i) {
         results[i] = run_single(circuit, observable, parameter_values[i]);
     }
